# Improve performance of opnorm for (m x 1) and (1 x n) sparse matrices