This PR improves the performance of `opnorm` for `m x 1` and `1 x n` sparse matrices by using `nzval` directly instead of allocating a dense array first.

To test the performance improvement, I have created two sparse matrices `A` and `B` with `A = sprandn(100_000, 1, 0.6)`, `B = sprandn(1, 100_000, 0.6)` and benchmarked `opnorm` using `BenchmarkTools.jl`. Here are the results:

| command        | Time (master) | Time (PR)   | Allocations (master)  | Allocations (PR)
| ------------- |-------------:| -----:|------:|-------:|
| `opnorm(A, 1)`  | 239.191 μs | 9.361 μs | 781.34 KiB | 16 bytes  |
| `opnorm(A)`  | 188.754 μs | 15.220 μs | 781.34 KiB | 16 bytes  |
| `opnorm(A, Inf)`  | 350.948 μs | 83.312 μs | 781.34 KiB | 16 bytes  |
| `opnorm(B, 1)`  |834.196 μs | 83.252 μs | 781.34 KiB | 16 bytes  |
| `opnorm(B)`  | 675.194 μs | 15.216 μs | 781.34 KiB  | 16 bytes |
| `opnorm(B, Inf)`  | 723.402 μs |9.448 μs | 781.34 KiB | 16 bytes  |

There could be one issue: Resizing `A.nzval` and changing a value of `A.nzval` outside of the array's size does affect the result of `opnorm(A, p)`. If that should not happen, this PR needs some rework. In that case, I would be thankful for some tips.